Split DaysIndex into DaysContainer & DaysList
