+++ src/App.js
@@ -5,7 +5,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import Nav from './components/Nav'
 import Home from './components/Home'
 import BirdsContainer from './components/BirdsContainer'
-import DaysIndex from './components/DaysIndex'
+import DaysContainer from './components/DaysContainer'
 import { connect } from 'react-redux'
 import { fetchBirds } from './redux/actions/birdActions'
 import { fetchDays } from './redux/actions/dayActions'
@@ -25,7 +25,7 @@ class App extends React.Component {
           <Switch>
             <Route exact path='/' component={Home} />
             <Route path='/birds' component={BirdsContainer} />
-            <Route path='/days' component={DaysIndex} />
+            <Route path='/days' component={DaysContainer} />
           </Switch>
         </Router>
       </div>
